requirements.txt
fastapi
uvicorn
requests
python-dotenv
azure-identity
azure-keyvault-secrets
azure-storage-blob
azure-servicebus
azure-functions
pytz
azure-search-documents==11.4.0


